ImGui sample: optimize handling of index/vertex data array
Ensure that we (re)allocate combined index and vertex data arrays
only when necessary - i.e., first allocation and when number of
elements changes. Either of these scenarios imply that the data
has changed, too.

If the buffer size has not changed, check each chunk (corresponding
to all data from a command list) for change, and overwrite it if
it has changed. Once a change is detected, skip the checks and
overwrite all subsequent data.

This new version of change detection allows us to populate the
array as necessary and detect change (so that we upload the
array to GPU buffer) while keeping only one instance of the
array.