Cross-platform sample: supply fallback surface dimensions
When calling Initialize and Resize on SceneView that was created
with Surface, supply fallback dimensions as obtained from the
presentation control.

This ensures that the SceneView is correctly (re)sized when the
surface itself does not provide the size information (i.e., its
current extent has width and height set to uint.MAX_VALUE); this
happens, for example, with Silk.NET under Linux with Wayland.